Rename component files as per their component class name
