+++ client/zestyreactnative/src/components/message/message.js
