+++ client/zestyreactnative/src/components/message/messageList.js
@@ -5,7 +5,7 @@ import {
   Text,
   View
 } from 'react-native';
-import Message from '../message/view';
+import Message from './message';
 import { border } from '../../helpers/scaffolding';
 import { getMessages } from '../../helpers/api';
